#156 [mod] : 타입 변경

+++ app/src/main/java/com/byeboo/app/data/mapper/quest/QuestStateMapper.kt
@@ -9,7 +9,7 @@ fun QuestCountResponseDto.toDomain(): QuestStateModel {
     return QuestStateModel(
         todayComplete = this.todayComplete,
         userCurrentStatus = this.userCurrentStatus,
-        count = this.count.toInt()
+        count = this.count.toLong()
     )
 }
